+++ src/Consumable.php
@@ -36,6 +36,7 @@
 use Glpi\Application\View\TemplateRenderer;
 use Glpi\DBAL\QuerySubQuery;
 use Glpi\Event;
+use Glpi\Search\SearchEngine;
 
 //!  Consumable Class
 /**
@@ -69,6 +70,16 @@ class Consumable extends CommonDBChild
 
         $forbidden   = parent::getForbiddenStandardMassiveAction();
         $forbidden[] = 'update';
+        $forbidden[] = 'ObjectLock:unlock';
+        $forbidden[] = 'add_note';
+        $forbidden[] = 'add_transfer_list';
+
+        // Despite using the Clonable trait, the 'clone' option was not available
+        // in the massive actions defined by the old Consumable::showForConsumableItem()
+        // method.
+        // To keep things consistent, clone is blacklisted here.
+        $forbidden[] = 'clone';
+
         return $forbidden;
     }
 
@@ -191,6 +202,21 @@ class Consumable extends CommonDBChild
         return false;
     }
 
+    public static function getMassiveActionsForItemtype(
+        array &$actions,
+        $itemtype,
+        $is_deleted = 0,
+        CommonDBTM $checkitem = null
+    ) {
+        // Special actions only for self
+        if ($itemtype != self::getType()) {
+            return;
+        }
+
+        $action_prefix = self::getType() . MassiveAction::CLASS_ACTION_SEPARATOR;
+        $actions[$action_prefix . 'backtostock'] = __('Back to stock');
+        $actions[$action_prefix . 'give'] = _x('button', 'Give');
+    }
 
     public static function showMassiveActionsSubForm(MassiveAction $ma)
     {
@@ -199,22 +225,21 @@ class Consumable extends CommonDBChild
         $input = $ma->getInput();
         switch ($ma->getAction()) {
             case 'give':
-                if (isset($input["entities_id"])) {
-                    Dropdown::showSelectItemFromItemtypes(['itemtype_name'
-                                                              => 'give_itemtype',
-                        'items_id_name'
-                                                              => 'give_items_id',
-                        'entity_restrict'
-                                                              => $input["entities_id"],
-                        'itemtypes'
-                                                              => $CFG_GLPI["consumables_types"]
-                    ]);
-                    echo "<br><br>" . Html::submit(
-                        _x('button', 'Give'),
-                        ['name' => 'massiveaction']
-                    );
-                    return true;
-                }
+                // Guess entity from first item
+                $consumable = current($input['items'][self::getType()]);
+                $entities_id = $consumable->fields['entities_id'];
+
+                Dropdown::showSelectItemFromItemtypes([
+                    'itemtype_name'   => 'give_itemtype',
+                    'items_id_name'   => 'give_items_id',
+                    'entity_restrict' => $entities_id,
+                    'itemtypes'       => $CFG_GLPI["consumables_types"]
+                ]);
+                echo "<br><br>" . Html::submit(
+                    _x('button', 'Give'),
+                    ['name' => 'massiveaction']
+                );
+                return true;
         }
         return parent::showMassiveActionsSubForm($ma);
     }
@@ -505,9 +530,9 @@ class Consumable extends CommonDBChild
      **/
     public static function showAddForm(ConsumableItem $consitem)
     {
+        Toolbox::deprecated('Replaced by Consumable::displayConsumableList()');
 
         $ID = $consitem->getField('id');
-
         if (!$consitem->can($ID, UPDATE)) {
             return;
         }
@@ -531,7 +556,6 @@ class Consumable extends CommonDBChild
         }
     }
 
-
     /**
      * Print out the consumables of a defined type
      *
@@ -543,6 +567,8 @@ class Consumable extends CommonDBChild
     public static function showForConsumableItem(ConsumableItem $consitem, $show_old = 0)
     {
         global $DB;
+
+        Toolbox::deprecated("Replaced by Consumable::displayConsumableList()");
 
         $tID = $consitem->getField('id');
         if (!$consitem->can($tID, READ)) {
@@ -680,7 +706,61 @@ class Consumable extends CommonDBChild
         echo "</div>";
     }
 
+    /**
+     * Display a consumable list for a given consumable item
+     *
+     * @param ConsumableItem $parent Parent consumable item
+     *
+     * @return void
+     */
+    public static function displayConsumableList(ConsumableItem $parent): void
+    {
+        // Search criteria used to display lists (used and unused items)
+        $criteria_parent_consumable = [
+            'link'       => 'AND',
+            'field'      => 8, // Parent consumable
+            'searchtype' => 'equals',
+            'value'      => $parent->getID(),
+        ];
+        $criteria_unused = [
+            [
+                'link'       => 'AND',
+                'field'      => 5, // Date out
+                'searchtype' => 'empty',
+                'value'      => "0",
+            ],
+            $criteria_parent_consumable,
+        ];
+        $criteria_used = [
+            [
+                'link' => 'AND NOT',
+                'criteria' => $criteria_unused,
+            ],
+            $criteria_parent_consumable,
+        ];
 
+        // Count used and unused items
+        $count_unused = countElementsInTable(self::getTable(), [
+            'date_out' => 'NULL',
+            ConsumableItem::getForeignKeyField() => $parent->getID(),
+        ]);
+        $count_used = countElementsInTable(self::getTable(), [
+            'NOT' => ['date_out' => 'NULL'],
+            ConsumableItem::getForeignKeyField() => $parent->getID(),
+        ]);
+
+        $twig = TemplateRenderer::getInstance();
+        $twig->display('pages/assets/consumable_list.html.twig', [
+            'item'            => new self(),
+            'parent'          => $parent,
+            'itemtype'        => self::getType(),
+            'can_edit'        => $parent->canUpdateItem(),
+            'criteria_unused' => $criteria_unused,
+            'criteria_used'   => $criteria_used,
+            'count_unused'    => $count_unused,
+            'count_used'      => $count_used,
+        ]);
+    }
     public static function showForUser(User $user)
     {
         global $DB;
@@ -972,9 +1052,7 @@ class Consumable extends CommonDBChild
 
         switch ($item->getType()) {
             case 'ConsumableItem':
-                self::showAddForm($item);
-                self::showForConsumableItem($item);
-                self::showForConsumableItem($item, 1);
+                self::displayConsumableList($item);
                 break;
             case 'User':
                 self::showForUser($item);
@@ -1013,5 +1091,116 @@ class Consumable extends CommonDBChild
         }
 
         return $sql_filters;
+    }
+
+    public function rawSearchOptions()
+    {
+        $options = parent::rawSearchOptions();
+
+        $options[] = [
+            'id'                 => '2',
+            'table'              => $this->getTable(),
+            'field'              => 'id',
+            'name'               => __('ID'),
+            'massiveaction'      => false,
+            'datatype'           => 'number'
+        ];
+
+        $options[] = [
+            'id'                 => '3',
+            'table'              => $this->getTable(),
+            'field'              => 'date_out',
+            'name'               => _n('State', 'States', 1),
+            'massiveaction'      => false,
+            'nosearch'           => true,
+            'datatype'           => 'specific',
+        ];
+
+        $options[] = [
+            'id'                 => '4',
+            'table'              => $this->getTable(),
+            'field'              => 'date_in',
+            'name'               => __('Add date'),
+            'massiveaction'      => false,
+            'datatype'           => 'date'
+        ];
+
+        $options[] = [
+            'id'                 => '5',
+            'table'              => $this->getTable(),
+            'field'              => 'date_out',
+            'name'               => __('Use date'),
+            'massiveaction'      => false,
+            'datatype'           => 'date'
+        ];
+
+        $options[] = [
+            'id'                 => '6',
+            'table'              => $this->getTable(),
+            'field'              => 'items_id',
+            'name'               => __('Given to'),
+            'massiveaction'      => false,
+            'additionalfields'   => ['itemtype'],
+            'datatype'           => 'specific',
+            'searchtype'         => ['equals', 'notequals'],
+        ];
+
+        $infocom_label = Infocom::getTypeName();
+        $options[] = [
+            'id'                 => '7',
+            'table'              => $this->getTable(),
+            'field'              => 'id',
+            'name'               => $infocom_label,
+            // Content will be very short so using an icon here save space
+            'htmlname'           => "<i class=\"fa fa-coins\" title=\"$infocom_label\"></i>",
+            'massiveaction'      => false,
+            'nosearch'           => true,
+            'datatype'           => 'specific',
+            'nosort'             => 'true',
+        ];
+
+        $options[] = [
+            'id'                 => '8',
+            'table'              => ConsumableItem::getTable(),
+            'field'              => 'name',
+            'name'               => ConsumableItem::getTypeName(1),
+            'massiveaction'      => false,
+            'datatype'           => 'dropdown',
+            'searchtype'         => ['equals', 'notequals'],
+        ];
+
+        return $options;
+    }
+
+    public static function getSpecificValueToDisplay(
+        $field,
+        $values,
+        array $options = []
+    ) {
+        switch ($options['searchopt']['id']) {
+            case '3': // State
+                $date_out = $values['date_out'];
+                return empty($date_out) ? __("New") : __("Used");
+
+            case '6': // Given to
+                $itemtype = $values['itemtype'];
+                $items_id = $values['items_id'];
+                if (!is_null($itemtype)) {
+                    $item = new $itemtype();
+                    $item->getFromDB($items_id);
+                    return $item->getLink();
+                }
+
+                // Must not be empty
+                return " ";
+
+            case '7': // Infocom shortcut
+                $id = $values['id'];
+                ob_start();
+                Infocom::showDisplayLink(Consumable::getType(), $id);
+                return ob_get_clean();
+        }
+
+        return parent::getSpecificValueToDisplay($field, $values, $options);
     }
 }