+++ CHANGELOG.md
@@ -51,6 +51,7 @@ The present file will list all changes made to the project; according to the
 
 ### Deprecated
 - Survey URL tags `TICKETCATEGORY_ID` and `TICKETCATEGORY_NAME` are deprecated and replaced by `ITILCATEGORY_ID` and `ITILCATEGORY_NAME` respectively.
+- `Consumable::showAddForm()` and `Consumable::showForConsumableItem()` are deprecated and replaced by `Consumable::displayConsumableList()`.
 
 ### Removed
 - XML-RPC API.